Changement type scanPartition
Auparavant la fonction renvoyait un long, maintenant un int.

+++ MonteCarloMethod/schurNumberNestedMonteCarloHeader.h
@@ -38,7 +38,7 @@ void partition_copy(partition_t *partitionstrucd, partition_t *partitionstrucs);
 
 /*Fonctions d'affichage*/
 void printPartition(unsigned int p, unsigned long n, mp_limb_t **partition);
-unsigned long schurNumberScanPartitionFromFile(char *filename, partition_t *partitionstruc);
+unsigned int schurNumberScanPartitionFromFile(char *filename, partition_t *partitionstruc);
 
 /*Fonction ajoutant les entiers un par un*/
 unsigned long schurNumberSimpleNestedMonteCarlo(unsigned int p, unsigned long *narray, unsigned int level, unsigned int simulnum,
